win64: misc pointer truncation fixes

+++ game/server/ai_navigator.cpp
@@ -1222,7 +1222,7 @@ float CAI_Navigator::GetPathTimeToGoal()
 AI_PathNode_t CAI_Navigator::GetNearestNode()
 {
 #ifdef WIN32
-	COMPILE_TIME_ASSERT( (int)AIN_NO_NODE == NO_NODE );
+	COMPILE_TIME_ASSERT( (intp)AIN_NO_NODE == NO_NODE );
 #endif
 	return (AI_PathNode_t)(intp)( GetPathfinder()->NearestNodeToNPC() );
 }
